Refactor RetrieveSourceDocuments into SourceDocumentDetails page model
