+++ src/Portal/Portal/Pages/DbAssessment/Review.cshtml
@@ -5,6 +5,7 @@
 }
 
 <script type="text/javascript" src="~/js/Review.js"></script>
+<script type="text/javascript" src="~/js/_SourceDocumentDetails.js"></script>
 
 <div class="row mt-1">
     <div class="col-4">
