+++ src/Portal/Portal/Pages/DbAssessment/_SourceDocumentDetails.cshtml
@@ -1,8 +1,11 @@
-﻿@using WorkflowDatabase.EF
+﻿@page
+@using WorkflowDatabase.EF
+
 @model Portal.Pages.DbAssessment._SourceDocumentDetailsModel
 @*
     For more information on enabling MVC for empty projects, visit http://go.microsoft.com/fwlink/?LinkID=397860
 *@
+@{ Layout = null; }
 <div class="card mb-4">
     <h6>Source Document Details</h6>
     <div class="card-body">
